tests/new_indicators_addition_test.py
import unittest
import pandas as pd
from logic import calculate_indicators, import_data
import numpy as np
import os

class TestCalculateIndicators(unittest.TestCase):
	@classmethod
	def setUpClass(cls):
		# Create a sample CSV file with mock data
		cls.test_csv = 'test_data.csv'
		data = {
			'date': pd.date_range(start='2021-01-01', periods=60, freq='D').astype(np.int64) // 10**6,
			'Open': range(1, 61),
			'High': range(2, 62),
			'Low': range(1, 61),
			'Close': range(2, 62),
			'Volume': range(100, 160)
		}
		df = pd.DataFrame(data)
		df.to_csv(cls.test_csv, index=False)
		df = import_data(cls.test_csv)
		df.to_csv(cls.test_csv, index=False)

	@classmethod
	def tearDownClass(cls):
		# Cleanup: Remove the sample CSV file
		os.remove(cls.test_csv)
		pass

	def test_empty_dataframe(self):
		# Modify the test to create a CSV file with only column headers
		df = pd.DataFrame(columns=['date', 'Open', 'High', 'Low', 'Close', 'Volume'])
		empty_csv = "test_data_empty.csv"
		df.to_csv(empty_csv, index=False)
		result_df = calculate_indicators(empty_csv)
		os.remove(empty_csv)
		expected_columns = ['date', 'Open', 'High', 'Low', 'Close', 'Volume', 'SMA_10', 'SMA_50', 'EMA_20', 'RSI_14', 'MACD', 'BB_High', 'BB_Low', 'ATR_14']
		self.assertTrue(all(column in result_df.columns for column in expected_columns), "All expected columns should be present in the result DataFrame")
		self.assertEqual(len(result_df), 0, "Result DataFrame should be empty for empty input")

	def test_indicator_calculations(self):
		# Test that indicators are calculated correctly
		# This test would require known expected values for comparison, which are not provided here.
		# It's recommended to calculate these expected values manually or using a trusted tool, then compare.
		pass  # Placeholder for actual test implementation

	def test_output_structure(self):
		# Test that the output DataFrame has the correct structure
		result_df = calculate_indicators(self.test_csv)
		expected_order = ['date', 'Open', 'High', 'Low', 'Close', 'Volume', 'SMA_10', 'SMA_50', 'EMA_20', 'RSI_14', 'MACD', 'BB_High', 'BB_Low', 'ATR_14']
		self.assertEqual(list(result_df.columns), expected_order, "Columns are not in the expected order")

if __name__ == '__main__':
	unittest.main()